Add utility command tpr (Teleport Request)
A command to allow players to send a request to teleport to another player and allowing that other player to accept or deny the request. If no response is given then the request is canceled. While the request is pending other players cannot send a request to that specific player pending a response.